fix(audit): tie the format version to what reaches the wire
Bump checker:
- A fixture value change reports `values` and defers with guidance naming
  both readings; a fixture rename reports `unknown`.
- A removed `action_name` reports `breaking`, so renaming an action needs a
  MAJOR bump. Added actions leave the format unchanged. The self-test
  cross-checks the parsed names against the committed OpenAPI document.
- Git failures raise; only "no match" reads as an absent declaration.

Contract rules:
- Entity and action objects are read at their real positions, so
  client-supplied property keys are data and per-decision entries are
  checked.
- `contract` owns the failure-reason wire tags, and the definitive-denial
  set is derived from the enum by an exhaustive match.
- Ten negative tests assert the exact violation each rule reports.

CI:
- A tree-wide grep requires exactly one audit emission site.
- The release-notes gate reads the title and body from the API, and its
  section detector accepts `###` subheadings.

The audit test module lives in `audit/tests.rs`,
`GrantContextValue::size_hint` is exact, and the documentation check is
driven by `ENVELOPE_KEYS`.
